Updated prize core data model, added requestPassphraseAlertView UIViewController extension, fixed BananoQuest class, added quest data verification, fixed new wallet keyboard not dismissing.

+++ banano-quest/banano-quest/Models/Quest+CoreDataProperties.swift
@@ -24,7 +24,7 @@ extension Quest {
     @NSManaged public var merkleRoot: String?
     @NSManaged public var maxWinners: Int16
     @NSManaged public var questID: Int32
-    @NSManaged public var prize: Int32
+    @NSManaged public var prize: Double
     @NSManaged public var winners: Winners?
     @NSManaged public var metadata: Metadata?
